Fully Functional R Adapter and Edit to POST Payload File Saving
The zip file of output files is sent as raw binary in the HTTP response 'body' from the R adapter. Up to client to interpret 'body' as a zip file.

Also, all files sent in with the POST request are saved in binary. Delete 'bin.file.exts' (spelling?) field in 'Adapter' RefClass;